+++ src/app/lib/widgets/item-control/item-control.component.ts
@@ -5,6 +5,7 @@ import {FormService} from '../../../services/form.service';
 import {Subscription} from 'rxjs';
 import fhir from 'fhir/r4';
 import {Util} from '../../util';
+import {LiveAnnouncer} from "@angular/cdk/a11y";
 
 @Component({
   selector: 'lfb-item-control',
@@ -26,25 +27,63 @@ export class ItemControlComponent extends LfbControlWidgetComponent implements O
   isRepeat = false;
   answerMethod = 'answer-option';
   subscriptions: Subscription [] = [];
+  dataType;
 
-  constructor(private extensionsService: ExtensionsService, private formService: FormService, private cdr: ChangeDetectorRef) {
+  constructor(private extensionsService: ExtensionsService, private formService: FormService,
+              private cdr: ChangeDetectorRef, private liveAnnouncer: LiveAnnouncer) {
     super();
   }
 
+  /**
+   * Angular life cycle event - Initialize attributes.
+   */
   ngOnInit() {
     super.ngOnInit();
     this.init();
   }
 
   /**
+   * Compose a Group Item Control object from the schema.
+   */
+  composeGroupItemControlObject() {
+    if (this.formProperty?.schema?.oneOf) {
+      this.formProperty.schema.oneOf.forEach((groupItemControl) => {
+        this.optionsObj[groupItemControl.enum[0]] = groupItemControl.display;
+      });
+    }
+  }
+
+  /**
+   * Get the default item control based on the selected data type, or the item control defined in the
+   * extension.
+   * @param dataTypeChanged - indicates if there is a change to the data type. True if the
+   *                          data type changed; otherwise, False.
+   * @returns - item control
+   */
+  getItemControl(dataTypeChanged: boolean = false): string {
+    const ext = this.getItemControlExtension();
+    const defaultItemControl = (this.dataType === 'group') ? '' : 'drop-down';
+    if (dataTypeChanged)
+      return defaultItemControl;
+
+    return ext ? ext.valueCodeableConcept.coding[0].code : defaultItemControl;
+  }
+
+  /**
    * Read formProperty values.
    */
   init() {
-    const ext = this.getItemControlExtension();
-    this.option = ext ? ext.valueCodeableConcept.coding[0].code : 'drop-down';
+    this.dataType = this.formProperty.searchProperty('/type').value;
+    this.option = this.getItemControl(false);
     this.isRepeat = !!this.formProperty.searchProperty('/repeats').value;
     this.answerMethod = this.formProperty.searchProperty('/__$answerOptionMethods').value;
+
+    this.composeGroupItemControlObject();
   }
+
+  /**
+   * Setup subscriptions.
+   */
   ngAfterViewInit() {
     super.ngAfterViewInit();
 
@@ -57,9 +96,14 @@ export class ItemControlComponent extends LfbControlWidgetComponent implements O
     this.subscriptions.push(sub);
 
     sub = this.formProperty.searchProperty('/type').valueChanges.subscribe((type) => {
+      const changed = !(this.dataType === type);
+      this.dataType = type;
       // If type is not choice, cleanup the extension.
-      if (type !== 'choice' && type !== 'open-choice') {
+      if (type !== 'choice' && type !== 'open-choice' && type !== 'group') {
         this.extensionsService.removeExtensionsByUrl(ItemControlComponent.itemControlUrl);
+      } else {
+        this.option = this.getItemControl(changed);
+        this.updateItemControlExt(this.option);
       }
       this.cdr.markForCheck();
     })
@@ -72,7 +116,6 @@ export class ItemControlComponent extends LfbControlWidgetComponent implements O
       this.cdr.markForCheck();
     })
     this.subscriptions.push(sub);
-
   }
 
   /**
@@ -101,13 +144,17 @@ export class ItemControlComponent extends LfbControlWidgetComponent implements O
     else {
       this.option = option;
     }
+
     const ext = this.getItemControlExtension();
-    if(!ext) {
-      this.extensionsService.addExtension(this.createExtension(option), 'valueCodeableConcept');
-    }
-    else {
-      ext.valueCodeableConcept.coding[0].code = option;
-      ext.valueCodeableConcept.coding[0].display = this.optionsObj[option];
+    if (option) {
+      if(!ext) {
+        this.extensionsService.addExtension(this.createExtension(option), 'valueCodeableConcept');
+      }
+      else {
+        delete ext.valueCodeableConcept.text;
+        ext.valueCodeableConcept.coding[0].code = option;
+        ext.valueCodeableConcept.coding[0].display = this.optionsObj[option];
+      }
     }
   }
 
@@ -147,9 +194,33 @@ export class ItemControlComponent extends LfbControlWidgetComponent implements O
     }
   }
 
+  /**
+   * Remove subscriptions before removing the component.
+   */
   ngOnDestroy() {
     this.subscriptions.forEach((sub) => {
       sub.unsubscribe();
     })
+  }
+
+  /**
+   * Compose the Group item control label to be announced by the screen reader.
+   * @param opt - JSON schema
+   * @returns - text to be read by the screen reader.
+   */
+  composeGroupItemControlLabel(opt: any): string {
+    let label = `Group item control ${opt.display}. ${opt.description}  `;
+    if (!opt.support)
+      label += "Please note that this item control is not supported by the LHC-Forms preview.";
+    return label;
+  }
+
+  /**
+   * Clear selection for the 'Group Item Control' radio button.
+   */
+  clearGroupItemControlSelection() {
+    this.option = '';
+    this.extensionsService.removeExtensionsByUrl(ItemControlComponent.itemControlUrl);
+    this.liveAnnouncer.announce('Group item control selection has been cleared.');
   }
 }