+++ src/app/lib/widgets/observation-link-period/observation-link-period.component.ts
@@ -12,7 +12,6 @@ import {ExtensionsService} from '../../../services/extensions.service';
 import {StringComponent} from '../string/string.component';
 import {fhirPrimitives} from '../../../fhir';
 
-
 interface ObservationLinkPeriodExtension {
   url: fhirPrimitives.url,
   valueDuration: {