+++ src/app/lib/widgets/string/string.component.ts
@@ -1,14 +1,17 @@
 /**
  * Component for general input box
  */
-import {AfterViewInit, Component, OnInit} from '@angular/core';
+import {AfterViewInit, Component, inject, OnInit} from '@angular/core';
 import {LfbControlWidgetComponent} from '../lfb-control-widget/lfb-control-widget.component';
+import { LiveAnnouncer } from '@angular/cdk/a11y';
 
 @Component({
   selector: 'lfb-string',
   templateUrl: './string.component.html'
 })
 export class StringComponent extends LfbControlWidgetComponent implements OnInit, AfterViewInit {
+
+  liveAnnouncer = inject(LiveAnnouncer);
 
   // Replace standard error messages from schema validator with customized messages.
   // Keys are error codes from the validator.
@@ -37,6 +40,11 @@ export class StringComponent extends LfbControlWidgetComponent implements OnInit
   errors: {code: string, originalMessage: string, modifiedMessage: string} [] = null;
 
   Array = Array; // To use in templates.
+
+  constructor() {
+    super();
+  }
+
   ngOnInit() {
     super.ngOnInit();
     this.controlClasses = this.controlClasses || 'form-control form-control-sm';
@@ -77,5 +85,16 @@ export class StringComponent extends LfbControlWidgetComponent implements OnInit
       return el.pattern === pattern;
     });
     return messageObj ? messageObj.message : null;
+  }
+
+
+  /**
+   * Check for errors when the field is focused and announce any existing errors.
+   */
+  announceErrors(): void {
+    if(this.errors) {
+      const combinedErrorMessage = this.errors.reduce((acc, error) => acc + error.originalMessage, '');
+      this.liveAnnouncer.announce(combinedErrorMessage);
+    }
   }
 }